+++ packages/jest-cli/src/cli/index.js
@@ -15,6 +15,7 @@ import {
   Console,
   clearLine,
   createDirectory,
+  realpath,
   validateCLIOptions,
 } from 'jest-util';
 import {readConfig} from 'jest-config';
@@ -143,8 +144,7 @@ const getProjectListFromCLIArgs = (argv, project: ?Path) => {
 
   if (!projects.length && process.platform === 'win32') {
     try {
-      // $FlowFixMe
-      projects.push(process.binding('fs').realpath(process.cwd()));
+      projects.push(realpath(process.cwd()));
     } catch (err) {
       // do nothing, just catch error
       // process.binding('fs').realpath can throw, e.g. on mapped drives